Feer/web/backend/feerweb/requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.5
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.7
django-import-export==2.2.0
django-widget-tweaks==1.4.8
et-xmlfile==1.0.1
future==0.18.2
importlib-metadata==1.6.1
jdcal==1.4.1
kombu==4.6.10
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.2
pkg-resources==0.0.0
psycopg2==2.8.5
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
sqlparse==0.3.1
tablib==2.0.0
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
